pyproject.toml
[project]
name = "bbblb"
dynamic = ["version"]
description = "Yet another load balancer for BigBlueButton"
readme = "README.md"
authors = [
  {name = "Marcel Hellkamp", email = "me@defnull.de"},
]
license = "AGPL-3.0-or-later"
license-files = ["LICEN[CS]E*"]

requires-python = ">=3.13"
dependencies = [
    "aiodns>=3.5.0",
    "aiohttp>=3.13.1",
    "aiosqlite>=0.21.0",
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "click>=8.3.0",
    "lxml>=6.0.2",
    "psycopg[binary]>=3.2.12",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
    "starlette>=0.48.0",
]

[project.optional-dependencies]
server = [
    "httptools>=0.7.1",
    "uvicorn>=0.38.0",
    "uvloop>=0.22.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "lxml-stubs>=0.5.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "uvicorn>=0.38.0",
]
docs = [
    "m2r2>=0.3.4",
    "sphinx>=8.2.3",
    "sphinx-autobuild>=2025.8.25",
    "sphinx-click>=6.1.0",
    "watchdog>=6.0.0",
]

[project.scripts]
bbblb = "bbblb.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "bbblb/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
  "bbblb/",
  "/uv.lock",
  "docs/"
]
exclude = [

]

[tool.hatch.build.targets.wheel]
packages = ["bbblb"]
